backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mataLmod-backend"
version = "0.1.0"
description = "MaTaLmod FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.2",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "pydantic[email]>=2.7.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "anthropic>=0.28.0",
    "redis[hiredis]>=5.0.4",
    "celery[redis]>=5.4.0",
    "flower>=2.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    # Rate limiting (global API throttle)
    "slowapi>=0.1.9",
    # Observability
    "sentry-sdk[fastapi]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
